requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor==3.0.0
pillow
psutil
telegraph
py-tgcalls==0.9.7
pykeyboard
pyrogram==2.0.106
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
numpy
pytz
dnspython
youtube-search
youtube-search-python
git+https://github.com/yt-dlp/yt-dlp@master
git+https://github.com/alexmercerind/youtube-search-python@main
git+https://github.com/joetats/youtube_search@master
